Don't depend on easyjson
This library added support for easyjson (github.com/mailru/easyjson) in
January. Great!

But for many users (e.g. us) using just encoding/json is fine, but this
repo now depends on github.com/mailru/easyjson only because of a few
references to their Marshal/Unmarshal interfaces.

This change copies those interfaces to the library in `easyjson.go`, so
that projects that add this library as a dependency and don't use
easyjson no longer needlessly add `easyjson` in vendor.